testrecorder/requirements.txt
asgiref==3.5.0
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Django==2.2.7
djangorestframework==3.13.1
ffmpeg-python==0.2.0
future==0.18.2
google-api-core==2.7.2
google-api-python-client==2.44.0
google-auth==2.6.4
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
mega.py==1.0.8
oauthlib==3.2.0
pathlib==1.0.1
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
pyOpenSSL==22.0.0
pyparsing==3.0.8
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tenacity==5.1.5
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
zope.interface==5.4.0
